+++ app/controllers/TestCategoryController.php
@@ -53,7 +53,7 @@ class TestCategoryController extends \BaseController {
 			try{
 				$testcategory->save();
 				return Redirect::route('testcategory.index')
-							->with('message', 'messages.success-creating-test-category');
+					->with('message', trans('messages.success-creating-test-category'));
 			}catch(QueryException $e){
 				Log::error($e);
 			}
@@ -112,7 +112,8 @@ class TestCategoryController extends \BaseController {
 			$testcategory->save();
 
 			// redirect
-			return Redirect::route('testcategory.index')->with('message', 'messages.success-updating-test-category');
+			return Redirect::route('testcategory.index')
+				->with('message', trans('messages.success-updating-test-category'));
 		}
 	}
 
@@ -138,11 +139,23 @@ class TestCategoryController extends \BaseController {
 		//Soft delete the test category
 		$testcategory = TestCategory::find($id);
 
-		$testcategory->delete();
-
+		$testCategoryInUse = TestType::where('test_category_id', '=', $id)->first();
+		if (empty($testCategoryInUse)) {
+		    // The test category is not in use
+			$testcategory->delete();
+		} else {
+		    // The test category is in use
+		    return Redirect::route('testcategory.index')
+		    	->with('message', trans('messages.failure-test-category-in-use'));
+		}
 		// redirect
-		return Redirect::route('testcategory.index')->with('message', 'messages.success-deleting-test-category');
+		return Redirect::route('testcategory.index')
+			->with('message', trans('messages.success-deleting-test-category'));
 	}
-
-
 }
+
+
+
+
+
+
